+++ src/Benchmarking/DijkstraBenchmark.h
@@ -17,8 +17,8 @@ using namespace std;
 class DijkstraBenchmark {
 public:
     static void runAndMeasure(const vector < pair < unsigned int, unsigned int> > & trips, const Graph & graph);
-    static void runAndMeasureWithOutput(const vector < pair < unsigned int, unsigned int> > & trips, const Graph & graph, vector < long long unsigned int > & distances);
-    static double runAndMeasureOutputAndRetval(const vector < pair < unsigned int, unsigned int> > & trips, const Graph & graph, vector < long long unsigned int > & distances);
+    static void runAndMeasureWithOutput(const vector < pair < unsigned int, unsigned int> > & trips, const Graph & graph, vector < unsigned int > & distances);
+    static double runAndMeasureOutputAndRetval(const vector < pair < unsigned int, unsigned int> > & trips, const Graph & graph, vector < unsigned int > & distances);
 };
